Update README.md
Use all the skills you've learned up until now to answer the following questions as well as possible. 

1. What are the most dangerous types of sharks to humans? 
2. Are children more likely to be attacked by sharks? 
3. Are shark attacks where sharks were provoked more or less dangerous? 
4. Are certain activities more likely to result in a shark attack? 

If you feel you can't answer a question based on the dataset alone, feel free to find other datasets and use them in answering the questions.

For each answer you give not only answer the question but also write about the assumptions you made in answering the question. If an assumption or decision possibly created a bias please write about this as well.